Refactor namespace and enums.
